# api/routes/orders.js
const express = require('express');
const router = express.Router();

router.get('/', (req, res, next) => {
	res.status(200).json({
		message: 'Orders were fetched'
	});
});

router.post('/', (req, res, next) => {
	const order = {
		productId: req.body.productId,
		quantity: req.body.quantity
	};
	res.status(201).json({
		message: 'Order was created',
		order: order
	});
});

router.get('/:orderID', (req, res, next) => {
	res.status(200).json({
		message: `A delete route for /order/${req.params.orderID}`
	});
});

router.delete('/:orderID', (req, res, next) => {
	res.status(200).json({
		message: `A delete route for /order/${req.params.orderID}`
	});
});

module.exports = router;